requirements.txt
PyPDF2
openai
anthropic
python-dotenv